Extract local PTY launch helpers
